+++ Scenes/Player/player.gd
@@ -25,12 +25,6 @@ var temp_ignore_bodies : Array = []
 var gravity_disabled : bool = false
 var jump_over_timeout : Tween
 
-#QTE variables:
-var qte_keys = ["W","A","S","D"]
-var current_key : String
-var prev_key    : String
-var qte_active  : int = 0
-
 func _ready() -> void:
 	jump_over_timeout = get_tree().create_tween()
 
@@ -107,7 +101,7 @@ func _physics_process(delta: float) -> void:
 		temp_ignore_bodies.clear()
 	
 	position = position + velocity * delta
-
+	
 func cancel_jump():
 	if gravity_disabled:
 		jump_over_timeout.stop()
@@ -161,35 +155,5 @@ func join_floor(floor : Floor):
 	change_player_state(PLAYER_STATE.ON_FLOOR)
 	current_scenery = floor
 
-#Prompt user for key, start countdown, enable checking for input
-func start_qte(key : String, time : float) :
 	
-	qte_active = 1
-	print("Press " + key + "!")
-	$TimerQTE.start(time)
-	
-	
-#Disable checking for input and signal result:
-func end_qte(passed: bool = false) :
-	
-	qte_active = 0
-	
-	if passed :
-		emit_signal("qte_passed")
-		print("QTE PASSED!\n")
-	else :
-		emit_signal("qte_failed")
-		print("QTE FAILED\n")
-	
-#End QTE on timeout:
-func _on_timer_qte_timeout() -> void:
-	end_qte()
-	
-#End QTE on input, check result:
-func _input(event) :
-	if qte_active ==1 and Input.is_anything_pressed() :
-		if event.is_action_pressed(current_key) :
-			end_qte(true)
-		else :
-			end_qte()